test(radio): migrate states
