+++ core/src/components/radio/test/states/radio.e2e.ts
@@ -1,0 +1,48 @@
+import { expect } from '@playwright/test';
+import { configs, test } from '@utils/test/playwright';
+
+configs({ directions: ['ltr'] }).forEach(({ title, screenshot, config }) => {
+  test.describe(title('radio: states'), () => {
+    test('should render disabled checked radio correctly', async ({ page }) => {
+      await page.setContent(
+        `
+        <ion-radio-group value="1">
+          <ion-radio disabled="true" value="1">Label</ion-radio>
+        </ion-radio-group>
+      `,
+        config
+      );
+
+      const radio = page.locator('ion-radio');
+      expect(await radio.screenshot()).toMatchSnapshot(screenshot(`radio-checked-disabled`));
+    });
+
+    test('should render checked radio correctly', async ({ page }) => {
+      await page.setContent(
+        `
+        <ion-radio-group value="true">
+          <ion-radio value="true">Label</ion-radio>
+        </ion-radio-group>
+      `,
+        config
+      );
+
+      const radio = page.locator('ion-radio');
+      expect(await radio.screenshot()).toMatchSnapshot(screenshot(`radio-checked`));
+    });
+
+    test('should render unchecked radio correctly', async ({ page }) => {
+      await page.setContent(
+        `
+        <ion-radio-group>
+          <ion-radio value="true">Label</ion-radio>
+        </ion-radio-group>
+      `,
+        config
+      );
+
+      const radio = page.locator('ion-radio');
+      expect(await radio.screenshot()).toMatchSnapshot(screenshot(`radio-unchecked`));
+    });
+  });
+});
