+++ core/src/components/radio/test/states/radio.e2e.ts-snapshots/radio-checked-disabled-md-ltr-Mobile-Firefox-linux.png
