+++ core/src/components/radio/test/states/radio.e2e.ts-snapshots/radio-checked-ios-ltr-Mobile-Firefox-linux.png
