+++ core/src/components/radio/test/states/radio.e2e.ts-snapshots/radio-checked-ios-ltr-Mobile-Safari-linux.png
