+++ core/src/components/radio/test/states/radio.e2e.ts-snapshots/radio-checked-ios-rtl-Mobile-Chrome-linux.png
