+++ core/src/components/radio/test/states/radio.e2e.ts-snapshots/radio-checked-md-ltr-Mobile-Chrome-linux.png
