+++ core/src/components/radio/test/states/radio.e2e.ts-snapshots/radio-checked-md-rtl-Mobile-Firefox-linux.png
